Update hiredis submodule to 0.14.1
That fixes an issue on CentOS 7 which lead to lose the subscription
for a channel when a UNSUBSCRIBE is immediately followed by a SUBSRCRIBE. Then,
no listener is called when a PUBLISH is received for this channel. Furthermore,
Flexisip never send SUBSCRIBE again because it thinks that it is already subscribed,
so the problem remains until Flexisip is terminated.

This commit also add a test suite for RegistrarDb object.